Update 01_prepare_separate_generic_absa.R
This program has trouble loading does not work on line 23 because the path on line 9 does not exist. The https://github.com/Wesleyan-Media-Project/entity_linking/tree/main/facebook/data folder has 3 files similar to that of line 9 but I'm not sure which is correct. Also, those files are saved as .gz not .csv